Modularise code
